modified:   app/lib/dataProductos.ts 	modified:   app/lib/definitions.ts 	modified:   app/ui/infoProducto.tsx
